+++ urls.py
@@ -1,5 +1,3 @@
-
-
 class Urls:
 
     BASE_URL = 'https://qa-scooter.praktikum-services.ru/'
